+++ starsim/sim.py
@@ -65,12 +65,14 @@ class Sim(sc.prettyobj):
     @property
     def modules(self):
         # Return iterator over all Module instances (stored in standard places) in the Sim
+        products = [intv.product for intv in self.interventions.values() if hasattr(intv, 'product') and isinstance(intv.product, ss.Product)]
         return itertools.chain(
             self.demographics.values(),
             self.people.networks.values(),
             self.diseases.values(),
             self.connectors.values(),
             self.interventions.values(),
+            products,
             self.analyzers.values(),
         )
 
@@ -297,15 +299,23 @@ class Sim(sc.prettyobj):
                 errormsg = f'Intervention {intervention} does not seem to be a valid intervention: must be a function or Intervention subclass'
                 raise TypeError(errormsg)
 
-            for rng in intervention.rngs:
-                rng.initialize(self.rng_container, self.people.slot)
-
             # Add the intervention parameters and results into the Sim's dicts
             self.pars[intervention.name] = intervention.pars
             self.results[intervention.name] = intervention.results
 
             # Add intervention states to the People's dicts
             self.people.add_module(intervention)
+
+            # Intervention and product RNGs
+            for rng in intervention.rngs:
+                rng.initialize(self.rng_container, self.people.slot)
+
+            # If there's a product module present, initialize and add it
+            if hasattr(intervention, 'product') and isinstance(intervention.product, ss.Product):
+                intervention.product.initialize(self)
+                self.people.add_module(intervention.product)
+                for rng in intervention.product.rngs:
+                    rng.initialize(self.rng_container, self.people.slot)
 
         return