+++ CHANGELOG.rst
@@ -9,6 +9,13 @@ All notable changes to the codebase are documented in this file. Changes that ma
    :depth: 1
 
 
+Version 0.1.6 (2024-01-23)
+--------------------------
+- Adds template interventions and products for diagnostics and treatment
+- Adds syphilis screening & treatment interventions
+- *GitHub info*: PR `210 <https://github.com/amath-idm/stisim/pull/210>`_
+
+
 Version 0.1.5 (2024-01-23)
 --------------------------
 - Renamed ``stisim`` to ``starsim``.